+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 Release Notes
 =============
 
+## next
+
+- Add `AndroidUtils.isAtLeastUpsideDownCake`.
+
 ## 3.1.0
 _2022-12-09_